Port the branch classes
